# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Grove Street - Arabic Gaming Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    get:
      operationId: listGames
      tags: [products]
      summary: List games
      responses:
        "200":
          description: Games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createGame
      tags: [products]
      summary: Create game (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: gameId
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
...
  /rewards:
    get:
      operationId: listRewards
      tags: [points]
      summary: List available rewards
      responses:
        "200":
          description: Rewards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reward"
    post:
      operationId: createReward
      tags: [points]
      summary: Create reward (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
  /rewards/{id}/redeem:
    post:
      operationId: redeemReward
      tags: [points]
      summary: Redeem a reward
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Redeemed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedemptionResult"
  /ranks:
    get:
      operationId: listRanks
      tags: [points]
      summary: List ranks
      responses:
        "200":
          description: Ranks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rank"
    post:
      operationId: createRank
      tags: [points]
      summary: Create rank (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RankInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rank"
...
components:
  schemas:
...
    Game:
      type: object
      required: [id, name, nameAr, imageUrl, isActive]
      properties:
        id:       { type: integer }
        name:     { type: string }
        nameAr:   { type: string }
        imageUrl: { type: string }
        isActive: { type: boolean }
    GameInput:
      type: object
      required: [name, nameAr, imageUrl]
      properties:
        name:     { type: string }
        nameAr:   { type: string }
        imageUrl: { type: string }
        isActive: { type: boolean }
    Product:
      type: object
      required: [id, name, nameAr, price, category, isActive, stock, pointsEarned]
      properties:
        id:            { type: integer }
        name:          { type: string }
        nameAr:        { type: string }
        description:   { type: string }
        descriptionAr: { type: string }
        price:         { type: number }
        category:      { type: string }
        gameId:        { type: integer }
        gameName:      { type: string }
        imageUrl:      { type: string }
        isActive:      { type: boolean }
        isFeatured:    { type: boolean }
        stock:         { type: integer }
        pointsEarned:  { type: integer }
        inputFields:   { type: string }
    ProductInput:
      type: object
      required: [name, nameAr, price, category]
      properties:
        name:          { type: string }
        nameAr:        { type: string }
        description:   { type: string }
        descriptionAr: { type: string }
        price:         { type: number }
        category:      { type: string }
        gameId:        { type: integer }
        imageUrl:      { type: string }
        isActive:      { type: boolean }
        isFeatured:    { type: boolean }
        stock:         { type: integer }
        pointsEarned:  { type: integer }
        inputFields:   { type: string }
    ProductUpdate:
      type: object
      properties:
        name:          { type: string }
        nameAr:        { type: string }
        description:   { type: string }
        descriptionAr: { type: string }
        price:         { type: number }
        category:      { type: string }
        gameId:        { type: integer }
        imageUrl:      { type: string }
        isActive:      { type: boolean }
        isFeatured:    { type: boolean }
        stock:         { type: integer }
        pointsEarned:  { type: integer }
        inputFields:   { type: string }
...
    Reward:
      type: object
      required: [id, nameAr, pointsCost, isActive]
      properties:
        id:          { type: integer }
        nameAr:      { type: string }
        description: { type: string }
        pointsCost:  { type: integer }
        isActive:    { type: boolean }
        imageUrl:    { type: string }
    RewardInput:
      type: object
      required: [nameAr, pointsCost]
      properties:
        nameAr:      { type: string }
        description: { type: string }
        pointsCost:  { type: integer }
        isActive:    { type: boolean }
        imageUrl:    { type: string }
    RedemptionResult:
      type: object
      required: [success, newBalance, reward]
      properties:
        success:    { type: boolean }
        newBalance: { type: integer }
        reward:     { $ref: "#/components/schemas/Reward" }
        message:    { type: string }
    Rank:
      type: object
      required: [id, name, nameAr, minPoints, color, icon]
      properties:
        id:        { type: integer }
        name:      { type: string }
        nameAr:    { type: string }
        minPoints: { type: integer }
        color:     { type: string }
        icon:      { type: string }
        benefits:  { type: string }
    RankInput:
      type: object
      required: [name, nameAr, minPoints, color, icon]
      properties:
        name:      { type: string }
        nameAr:    { type: string }
        minPoints: { type: integer }
        color:     { type: string }
        icon:      { type: string }
        benefits:  { type: string }
...